pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "arrow"
authors = [{name = "Chris Smith", email = "crsmithdev@gmail.com"}]
readme = "README.rst"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-dateutil>=2.7.0",
    "backports.zoneinfo==0.2.1;python_version<'3.9'",
    "tzdata;python_version>='3.9'",
]
requires-python = ">=3.8"
description = "Better dates & times for Python"
keywords = [
    "arrow",
    "date",
    "time",
    "datetime",
    "timestamp",
    "timezone",
    "humanize",
]
dynamic = ["version"]

[project.optional-dependencies]
test = [
    "dateparser==1.*",
    "pre-commit",
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "pytz==2025.2",
    "simplejson==3.*",
]
doc = [
    "doc8",
    "sphinx>=7.0.0",
    "sphinx-autobuild",
    "sphinx-autodoc-typehints",
    "sphinx_rtd_theme>=1.3.0",
]

[project.urls]
Documentation = "https://arrow.readthedocs.io"
Source = "https://github.com/arrow-py/arrow"
Issues = "https://github.com/arrow-py/arrow/issues"

[tool.flit.module]
name = "arrow"
